NeighborlyAPI/requirements.txt
azure-functions==1.9.0
azure-cosmos
azure-cli-core
pymongo==4.0.2
pymongo[srv]
dnspython>=2.0.0